+++ agent_guard_core/cli.py
@@ -10,6 +10,7 @@ from mcp_proxy.config_loader import load_named_server_configs_from_file
 
 from agent_guard_core.config.config_manager import ConfigManager, ConfigurationOptions, SecretProviderOptions
 from agent_guard_core.proxy.audited_proxy import create_agent_guard_proxy_server
+from agent_guard_core.proxy.proxy_utils import get_audit_logger
 
 
 def get_cli_logger():
@@ -28,21 +29,6 @@ def get_cli_logger():
 get_cli_logger()
 
 
-def get_audit_logger() -> logging.Logger:
-    global formatter
-    audit_logger = logging.getLogger("agent_guard_core.audit")
-    audit_logger.setLevel(logging.DEBUG)
-    audit_file_handler = logging.FileHandler("agent_guard_core_proxy.log" if os.access(".", os.W_OK) else "/tmp/agent_guard_core_proxy.log")
-    audit_file_handler.setLevel(logging.DEBUG)
-    formatter = logging.Formatter('%(asctime)s %(levelname)s %(name)s: %(message)s')
-    audit_file_handler.setFormatter(formatter)
-    if not audit_logger.hasHandlers():
-        audit_logger.addHandler(audit_file_handler)
-
-    return audit_logger
-
-
-get_audit_logger()
 
 
 @click.group(help=(
@@ -65,11 +51,20 @@ def run():
     required=True,
     help="Path to the MCP server configuration file.",
 )
-def stdio_proxy(mcp_config_file):
+@click.option(
+    '--debug',
+    '-d','is_debug',
+    is_flag=True,
+    required=False,
+    default=False,
+    help="debug mode",
+
+)
+def stdio_proxy(mcp_config_file,is_debug):
     asyncio.run(_stdio_proxy_async(mcp_config_file))
 
 
-async def _stdio_proxy_async(mcp_config_file):
+async def _stdio_proxy_async(mcp_config_file,is_debug: bool = False):
     print(f"Starting stdio server from config  {mcp_config_file} ")
     base_env: dict[str, str] = {}
     stdio_params = load_named_server_configs_from_file(mcp_config_file, base_env)
@@ -78,13 +73,13 @@ async def _stdio_proxy_async(mcp_config_file):
 
     for name, params in stdio_params.items():
         logger.info(
-            "Setting up named server '%s': %s %s",
+            "Setting up mcp server '%s': %s %s",
             name,
             params.command,
             " ".join(params.args),
         )
     async with stdio_client(params, errlog=sys.stdout) as streams, ClientSession(*streams) as session:
-        app = await create_agent_guard_proxy_server(session, get_audit_logger())
+        app = await create_agent_guard_proxy_server(remote_app=session,logger=get_audit_logger(logging.DEBUG if is_debug else logging.INFO))
         async with stdio_server() as (read_stream, write_stream):
             await app.run(
                 read_stream,
@@ -128,10 +123,7 @@ default_provider = ConfigManager().get_config_value(
               '-ca',
               is_flag=True,
               help="Prompt for Conjur API Key.")
-@click.option('--target-mcp-server-config-file',
-              '-tc',
-              required=False,
-              help="Mcp server endpoint")
+
 def set(provider, conjur_authn_login, conjur_appliance_url, conjur_api_key):
     """
     Set the secret provider and related options in the Agent Guard configuration.
@@ -220,5 +212,5 @@ def get_param(key):
 cli.add_command(config)
 cli.add_command(run)
 
-if __name__ == '__main__':
-    cli(["run", "stdio-proxy", "--mcp-config-file", "config_example.json"], standalone_mode=False)
+# if __name__ == '__main__':
+#     cli(["run", "stdio-proxy", "--mcp-config-file", "config_example.json"], standalone_mode=False)
